+++ core/banking/erc20.go
@@ -64,7 +64,6 @@ func (e *Engine) EnableERC20(
 		bridgeView:  e.bridgeView,
 	}
 	e.assetActs[aa.id] = aa
-	e.bss.changedAssetActions = true
 	return e.witness.StartCheck(aa, e.onCheckDone, e.timeService.GetTimeNow().Add(defaultValidationDuration))
 }
 
@@ -163,7 +162,6 @@ func (e *Engine) ERC20WithdrawalEvent(
 
 	withd.WithdrawalDate = e.timeService.GetTimeNow().UnixNano()
 	withd.TxHash = txHash
-	e.bss.changedWithdrawals = true
 	e.broker.Send(events.NewWithdrawalEvent(ctx, *withd))
 
 	return nil