+++ core/staking/stake_verifier_snapshot.go
@@ -35,8 +35,6 @@ var (
 type stakeVerifierSnapshotState struct {
 	serialisedDeposited []byte
 	serialisedRemoved   []byte
-	changedDeposited    bool
-	changedRemoved      bool
 }
 
 func (s *StakeVerifier) serialisePendingSD() ([]byte, error) {
@@ -72,19 +70,12 @@ func (s *StakeVerifier) serialisePendingSR() ([]byte, error) {
 	return proto.Marshal(pl.IntoProto())
 }
 
-func (s *StakeVerifier) serialiseK(k string, serialFunc func() ([]byte, error), dataField *[]byte, changedField *bool) ([]byte, error) {
-	if !s.HasChanged(k) {
-		if dataField == nil {
-			return nil, nil
-		}
-		return *dataField, nil
-	}
+func (s *StakeVerifier) serialiseK(k string, serialFunc func() ([]byte, error), dataField *[]byte) ([]byte, error) {
 	data, err := serialFunc()
 	if err != nil {
 		return nil, err
 	}
 	*dataField = data
-	*changedField = false
 	return data, nil
 }
 
@@ -92,9 +83,9 @@ func (s *StakeVerifier) serialiseK(k string, serialFunc func() ([]byte, error), 
 func (s *StakeVerifier) serialise(k string) ([]byte, error) {
 	switch k {
 	case depositedKey:
-		return s.serialiseK(k, s.serialisePendingSD, &s.svss.serialisedDeposited, &s.svss.changedDeposited)
+		return s.serialiseK(k, s.serialisePendingSD, &s.svss.serialisedDeposited)
 	case removedKey:
-		return s.serialiseK(k, s.serialisePendingSR, &s.svss.serialisedRemoved, &s.svss.changedRemoved)
+		return s.serialiseK(k, s.serialisePendingSR, &s.svss.serialisedRemoved)
 	default:
 		return nil, types.ErrSnapshotKeyDoesNotExist
 	}
@@ -110,18 +101,6 @@ func (s *StakeVerifier) Keys() []string {
 
 func (s *StakeVerifier) Stopped() bool {
 	return false
-}
-
-func (s *StakeVerifier) HasChanged(k string) bool {
-	// switch k {
-	// case depositedKey:
-	// 	return s.svss.changedDeposited
-	// case removedKey:
-	// 	return s.svss.changedRemoved
-	// default:
-	// 	return false
-	// }
-	return true
 }
 
 func (s *StakeVerifier) GetState(k string) ([]byte, []types.StateProvider, error) {
@@ -167,7 +146,6 @@ func (s *StakeVerifier) restorePendingSD(ctx context.Context, deposited []*types
 		evts = append(evts, events.NewStakeLinking(ctx, *pending.IntoStakeLinking()))
 	}
 	var err error
-	s.svss.changedDeposited = false
 	s.svss.serialisedDeposited, err = proto.Marshal(p.IntoProto())
 	s.broker.SendBatch(evts)
 
@@ -217,7 +195,6 @@ func (s *StakeVerifier) restorePendingSR(ctx context.Context, removed []*types.S
 	}
 
 	var err error
-	s.svss.changedRemoved = false
 	s.svss.serialisedRemoved, err = proto.Marshal(p.IntoProto())
 	s.broker.SendBatch(evts)
 	return err