+++ core/types/mocks/restore_state_provider_mock.go
@@ -51,20 +51,6 @@ func (mr *MockPostRestoreMockRecorder) GetState(arg0 interface{}) *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetState", reflect.TypeOf((*MockPostRestore)(nil).GetState), arg0)
 }
 
-// HasChanged mocks base method.
-func (m *MockPostRestore) HasChanged(arg0 string) bool {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "HasChanged", arg0)
-	ret0, _ := ret[0].(bool)
-	return ret0
-}
-
-// HasChanged indicates an expected call of HasChanged.
-func (mr *MockPostRestoreMockRecorder) HasChanged(arg0 interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "HasChanged", reflect.TypeOf((*MockPostRestore)(nil).HasChanged), arg0)
-}
-
 // Keys mocks base method.
 func (m *MockPostRestore) Keys() []string {
 	m.ctrl.T.Helper()